07_epistatic_qtl CLI version

+++ scripts/07_epistatic_qtl.R
@@ -2,23 +2,56 @@ require(qtl)
 require(parallel)
 require(snow)
 require(dplyr)
-# setwd("~/projects/mdh_qtl")
 
-input_file <- "analyses/RIL_cross.csv"
-genotype <- c("A", "B")
+# Rscript scripts/07_epistatic_qtl.R RIL
+# Rscript scripts/07_epistatic_qtl.R B73_BC
+# Rscript scripts/07_epistatic_qtl.R Mo17_BC
+# Rscript scripts/07_epistatic_qtl.R B73_BC 100 6
+
+args <- commandArgs(trailingOnly = TRUE)
+population <- args[1]
+permutations <- ifelse(length(args) >= 2, as.numeric(args[2]), 100)
+cores <- ifelse(length(args) >= 3, as.numeric(args[3]), 4)
+
+presets <- list(
+  RIL = list(
+    input_file = "analyses/RIL_cross.csv",
+    genotype = c("A", "B"),
+    na.strings = "-",
+    crosstype = "ril",
+    phenos = 1
+  ),
+  B73_BC = list(
+    input_file = "analyses/B73_cross.csv",
+    genotype = c("AA", "AB"),
+    na.strings = "A-",
+    crosstype = "bc",
+    phenos = c(1, 5)
+  ),
+  Mo17_BC = list(
+    input_file = "analyses/Mo17_cross.csv",
+    genotype = c("AB", "BB"),
+    na.strings = "-B",
+    crosstype = "bc",
+    phenos = c(1, 5)
+  )
+)
+
+preset <- presets[[population]]
+
+input_file <- preset$input_file
+genotype <- preset$genotype
 alleles <- c("A", "B")
-na.strings <- "-"
-phenos <- 1
+na.strings <- preset$na.strings
+crosstype <- preset$crosstype
+phenos <- preset$phenos
 output_dir <- "analyses/qtl_analyses/"
-output_name <- "RIL"
-crosstype <- "ril"
-phenos <- c("NLB_WMD_BLUP")
-permutations <- 100
+output_name <- population
 
-ril=FALSE
-if(crosstype=="ril"){
-  ril=TRUE
-  crosstype="bc"
+ril <- FALSE
+if (crosstype == "ril") {
+  ril <- TRUE
+  crosstype <- "bc"
 }
 
 cross <- read.cross(format = "csv",
@@ -29,7 +62,7 @@ cross <- read.cross(format = "csv",
                     crosstype = crosstype)
 
 cross <- jittermap(cross)
-if(ril==TRUE){
+if (ril == TRUE) {
   cross <- convert2riself(cross)
 }
 
@@ -38,58 +71,41 @@ cross$pheno <- cross$pheno %>%
   mutate(across(everything(), as.character)) %>%
   mutate(across(everything(), as.numeric))
 
+# main scan (resumable)
+scantwo_temp_file <- paste0(output_dir, "tmp/", output_name, ".scantwo.tmp.RDS")
+if (file.exists(scantwo_temp_file)) {
+  out2 <- readRDS(scantwo_temp_file)
+} else {
+  out2 <- scantwo(cross, n.cluster = cores, pheno.col = phenos, verbose = TRUE)
+  saveRDS(out2, file = scantwo_temp_file)
+}
 
-cores = 4
-# permutations
-
-results_list <- list()
-
-
-# traits <- c("SLB_WMD_BLUE_MPH", "GLS_WMD_BLUE_MPH", "DTA_BLUE_MPH", "PH_BLUE_MPH", "EH_BLUE_MPH")
-pheno.cols <- match(phenos, colnames(cross$pheno))
-out2 <- scantwo(cross, n.cluster=cores, pheno.col = pheno.cols, verbose = TRUE)
-
-scantwo_temp_file <- paste(output_dir, "tmp/", output_name, ".scantwo.tmp.RDS", sep = "")
-saveRDS(out2, file = scantwo_temp_file)
-
-# permutations
-
+# permutations (resumable, batched by trait)
 batchnumber <- permutations / 10
 
-for(i in phenos){
-  pheno.col <- match(i, colnames(cross$pheno))
-  for(j in 1:batchnumber){
-    batch_filename <- paste(output_dir, "tmp/", output_name,".trait.", i, ".batch.", j, ".scantwo.perm.tmp", sep = "")
-    if(!file.exists(batch_filename)){
-      results <- scantwo(cross, n.cluster=cores,n.perm=10, pheno.col = i, verbose = TRUE)
+for (i in phenos) {
+  pheno_name <- colnames(cross$pheno)[i]
+  for (j in 1:batchnumber) {
+    batch_filename <- paste0(output_dir, "tmp/", output_name, ".trait.", pheno_name, ".batch.", j, ".scantwo.perm.tmp")
+    if (!file.exists(batch_filename)) {
+      results <- scantwo(cross, n.cluster = cores, n.perm = 10, pheno.col = i, verbose = TRUE)
       saveRDS(results, file = batch_filename)
-      print(paste(i, " batch ", j, " is done", sep=""))
+      print(paste(pheno_name, " batch ", j, " is done", sep = ""))
     }
   }
 }
 
-perm_list <- list()
-
-# read in temp files
-mqm_auto <- readRDS(mqm_temp_file)
-
+# combine permutation batches and save final output
 perm_results <- list()
-
-for(i in phenos){
-  pheno_name <- colnames(augmentedcross$pheno)[i]
-  for(j in 1:batchnumber){
-    batch_filename <- paste(output_dir, "tmp/", output_name,".trait.", pheno_name, ".batch.", j, ".perm.tmp", sep = "")
-    perm_list[[j]] <- readRDS(batch_filename)
-  }
-  processed_perms <- lapply(perm_list, mqmprocesspermutation)
-  combined_perms <- do.call(c, processed_perms)
-  perm_results[[i]] <- combined_perms
+for (i in phenos) {
+  pheno_name <- colnames(cross$pheno)[i]
+  perm_list <- lapply(1:batchnumber, function(j)
+    readRDS(paste0(output_dir, "tmp/", output_name, ".trait.", pheno_name, ".batch.", j, ".scantwo.perm.tmp")))
+  perm_results[[pheno_name]] <- Reduce(c, perm_list)
 }
 
+output <- list(scan = out2, permutations = perm_results)
 
-output <- list(scan = mqm_auto, permutations = perm_results)
-
-output_file <- paste(output_dir, output_name, ".RDS", sep="")
+output_file <- paste0(output_dir, output_name, "_scantwo.RDS")
 
 saveRDS(output, file = output_file)
-
